Postman updates
Updated the Postman collection to include new endpoints, a reorg of all the endpoints, and better descriptions on them as well. Also added a Postman environment file to go with the collection, and renamed the Docker Compose file to include a wp- prefix to fit in with the overall naming scheme I'm using for the Wordpress test demo project.